Rename suggestion views test file
